refactor(static methods): move fromHistory and fromLocation to separate files
